cloud_functions/daily_logs_requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-logging
google-cloud-storage
Flask<3.0
functions-framework==3.*